backend-fastapi/pyproject.toml
[project]
name = "live-english-tutor-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy==2.0.36",
    "psycopg2-binary==2.9.10",
    "pydantic-settings==2.6.0",
    "pydantic[email]==2.9.2",
    "firebase-admin>=6.5.0",
    "python-multipart==0.0.12",
    "httpx==0.27.2",
    "livekit-api==0.7.1",
]

[tool.uv]
package = false
